+++ backend/src/main/java/com/optimize/elykia/core/repository/DailyOperationLogRepository.java
@@ -2,6 +2,7 @@ package com.optimize.elykia.core.repository;
 
 import com.optimize.common.entities.repository.GenericRepository;
 import com.optimize.elykia.core.entity.report.DailyOperationLog;
+import com.optimize.elykia.core.enumaration.OperationType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
@@ -25,4 +26,18 @@ public interface DailyOperationLogRepository extends GenericRepository<DailyOper
 
         List<DailyOperationLog> findByDateBetweenAndCommercialUsername(LocalDate startDate, LocalDate endDate,
                         String commercialUsername);
+
+        Page<DailyOperationLog> findByDateBetweenAndType(LocalDate startDate, LocalDate endDate, OperationType type,
+                        Pageable pageable);
+
+        Page<DailyOperationLog> findByDateBetweenAndCommercialUsernameAndType(LocalDate startDate, LocalDate endDate,
+                        String commercialUsername, OperationType type, Pageable pageable);
+
+        Page<DailyOperationLog> findByDateAndCommercialUsernameAndType(LocalDate date, String commercialUsername,
+                        OperationType type, Pageable pageable);
+
+        List<DailyOperationLog> findByDateBetweenAndType(LocalDate startDate, LocalDate endDate, OperationType type);
+
+        List<DailyOperationLog> findByDateBetweenAndCommercialUsernameAndType(LocalDate startDate, LocalDate endDate,
+                        String commercialUsername, OperationType type);
 }